Prevent repository from replacing other valid files
The files copied to the system account carry the name of the
file in the original drive, not the name given in the repository.

We therefore cannot assume that the new file being added to the
system account is meant to replace a file already present.

This patch ensures that file names are always unique.